+++ include/integral/integral_special.hpp
@@ -6,12 +6,12 @@
 class IntegralSpecial : public NumericalMethod
 {
 public:
-    std::function<double(int)> functionToIntegrate;
+    std::function<double(double)> functionToIntegrate;
     double Xi;
     double Xf;
     double result;
 
-    IntegralSpecial(std::function<double(int)> function, double xi, double xf);
+    IntegralSpecial(std::function<double(double)> function, double xi, double xf);
 
     void accept(Visitor &visitor) const override;
