+++ include/integral/integral_special_simple_exponential_newton_cotes.hpp
@@ -1,15 +1,16 @@
 #pragma once
 #include "./integral_special.hpp"
 
-class IntegralSimpleExponentialNewtonCotes : public IntegralSpecial {
-    public:
-        double tolerance;
+class IntegralSimpleExponentialNewtonCotes : public IntegralSpecial
+{
+public:
+    double tolerance;
 
-        IntegralSimpleExponentialNewtonCotes(std::function<double(int)> function, double xi, double xf, double tolerance);
+    IntegralSimpleExponentialNewtonCotes(std::function<double(double)> function, double xi, double xf, double tolerance);
 
-        void execute() override;
+    void execute() override;
 
-        double Xs(double s) override;
+    double Xs(double s) override;
 
-        double DXs(double s) override;
+    double DXs(double s) override;
 };
