+++ src/integral/first_degree_integral_open_newtoncotes.cpp
@@ -1,11 +1,12 @@
 #include "integral/first_degree_integral_open_newtoncotes.hpp"
 #include "visitor/visitor.hpp"
 
-FirstDegreeIntegralOpenNewtonCotes::FirstDegreeIntegralOpenNewtonCotes(double xi, double deltax, std::function<double(int)> function, int numberOfPartitions) : Integral(xi, deltax, function, numberOfPartitions){};
+FirstDegreeIntegralOpenNewtonCotes::FirstDegreeIntegralOpenNewtonCotes(double xi, double deltax, std::function<double(double)> function, int numberOfPartitions) : Integral(xi, deltax, function, numberOfPartitions) {};
 
-FirstDegreeIntegralOpenNewtonCotes::FirstDegreeIntegralOpenNewtonCotes(double xi, double deltax, std::function<double(int)> function, double tolerance) : Integral(xi, deltax, function, tolerance){};
+FirstDegreeIntegralOpenNewtonCotes::FirstDegreeIntegralOpenNewtonCotes(double xi, double deltax, std::function<double(double)> function, double tolerance) : Integral(xi, deltax, function, tolerance) {};
 
-void FirstDegreeIntegralOpenNewtonCotes::accept(Visitor& visitor) const {
+void FirstDegreeIntegralOpenNewtonCotes::accept(Visitor &visitor) const
+{
     visitor.visit(*this);
 };
 
@@ -36,9 +37,12 @@ double FirstDegreeIntegralOpenNewtonCotes::calculateIntegralByError()
 
 void FirstDegreeIntegralOpenNewtonCotes::execute()
 {
-    if(this->numberOfPartitions != -1) {
+    if (this->numberOfPartitions != -1)
+    {
         this->result = this->calculateIntegralByPartitions();
-    }else {
+    }
+    else
+    {
         this->result = this->calculateIntegralByError();
     }
 };
