+++ src/integral/integral_gauss.cpp
@@ -1,12 +1,10 @@
 #include "integral/integral_gauss.hpp"
 #include "visitor/visitor.hpp"
 
-IntegralGauss::IntegralGauss(double xi, double xf, std::function<double(int)> function, vector<double> root, vector<double> weight)
-    : Xi{xi}, Xf{xf}, functionToIntegrate{function}, roots{root}, weights{weight} 
-    {};
+IntegralGauss::IntegralGauss(double xi, double xf, std::function<double(double)> function, vector<double> root, vector<double> weight)
+    : Xi{xi}, Xf{xf}, functionToIntegrate{function}, roots{root}, weights{weight} {};
 
-
-void IntegralGauss::accept(Visitor& visitor) const 
+void IntegralGauss::accept(Visitor &visitor) const
 {
     visitor.visit(*this);
 };
