+++ src/integral/integral_special.cpp
@@ -1,7 +1,7 @@
 #include "integral/integral_special.hpp"
 #include "visitor/visitor.hpp"
 
-IntegralSpecial::IntegralSpecial(std::function<double(int)> function, double xi, double xf)
+IntegralSpecial::IntegralSpecial(std::function<double(double)> function, double xi, double xf)
     : Xi{xi}, Xf{xf}, functionToIntegrate{function} {};
 
 void IntegralSpecial::accept(Visitor &visitor) const
