+++ src/integral/third_degree_integral_gausshermite.cpp
@@ -1,18 +1,16 @@
 #include "integral/third_degree_integral_gausshermite.hpp"
 #include "visitor/visitor.hpp"
 
-ThirdDegreeIntegralGuassHermite::ThirdDegreeIntegralGuassHermite(double xi, double xf, std::function<double(int)> function) : 
-    IntegralGauss(xi, xf, function, vector<double>{-1.224744871, 0, 1.224744871}, vector<double>{0.295408975, 1.181635901, 0.295408975}) 
-    {};
+ThirdDegreeIntegralGuassHermite::ThirdDegreeIntegralGuassHermite(double xi, double xf, std::function<double(double)> function) : IntegralGauss(xi, xf, function, vector<double>{-1.224744871, 0, 1.224744871}, vector<double>{0.295408975, 1.181635901, 0.295408975}) {};
 
 /**
  * This method is the accept method for the visitor pattern
  * used to get the result of the calculation of a method
  *
- * 
+ *
  * @return void - this method does not return any value, just prints the result.
  */
-void ThirdDegreeIntegralGuassHermite::accept(Visitor& visitor) const 
+void ThirdDegreeIntegralGuassHermite::accept(Visitor &visitor) const
 {
     visitor.visit(*this);
 };
@@ -22,17 +20,16 @@ double ThirdDegreeIntegralGuassHermite::getXk(int rootIndex)
     return this->roots[rootIndex];
 };
 
-
 /**
- * Used by the context class to call the execution of some numerical method 
+ * Used by the context class to call the execution of some numerical method
  * This method does not receive any parameter
- * 
- * @return void - This method does not return any value, just sets the result on the result attribute 
+ *
+ * @return void - This method does not return any value, just sets the result on the result attribute
  */
 void ThirdDegreeIntegralGuassHermite::execute()
 {
     double sum = 0;
-    for(int i = 0; i < roots.size(); ++i)
+    for (int i = 0; i < roots.size(); ++i)
     {
         sum += this->functionToIntegrate(this->getXk(i)) * this->weights[i];
     }
